move inference options to own file
